Client should always honor retry parameters.
It is reasonable to expect that fetching schema from transport honors
retry parameters. It currently does not do so, by circumventing the use
of the private `_get_result` method during instance construction.

This expectation is reasonable because connection issues are very likely
on first connection.

The difficulty is that self.transport is not yet resolved, for non-trivial
reasons, when we would like to use the private method. Therefore, proposed
solution involves parameterizing `transport` in the private method.