Add new exceptions (GQLSyntaxError and GQLServerError)
GQLServerError maps to errors which are generated on the backend server.
I.e. the server responds with a <200 OK> but an "errors" field is included
within the response. These are often user facing errors and should be
handled as such.

GQLSyntaxError maps to cases where either the query or the schema is
improperly formatted. This usually indicates a mistake in the code.