🏷 Add type overloads for `get_execution_result`
Methods like `execute` were previously annotated with a `Union` type,
indicating all of the types of return values that the function could be
returned. However, a more specific return type can often be inferred
statically based on whether or not the `get_execution_result` keyword
argument is passed to the method.

These changes add function type overloads to methods accepting the
`get_execution_result` keyword, which allows Mypy to infer a more
specific return type based on the passed value of
`get_execution_result`, if any.